Back api ressources

+++ app/Http/Controllers/API/RessourceController.php
@@ -8,34 +8,93 @@ use App\Http\Requests\Ressource\StoreUpdateRessourceRequest;
 use App\Http\Requests\Ressource\UpdateRessourceRequest;
 use App\Models\Commentaire;
 use App\Models\Jointure_ress_utilisateur;
+use App\Repositories\Favoris\FavorisRepository;
 use App\Repositories\LienRessourceRepository;
 use Illuminate\Http\Request;
 use App\Repositories\RessourceRepository;
-
+use Exception;
+use Illuminate\Support\Facades\Auth;
 
 class RessourceController extends Controller
 {
 
     private $ressourceRepository;
     private $lienRessourceRepository;
+    private $favorisRepository;
 
-    public function __construct(RessourceRepository $ressourceRepository, LienRessourceRepository $lienRessourceRepository)
+    public function __construct(RessourceRepository $ressourceRepository, LienRessourceRepository $lienRessourceRepository, FavorisRepository $favorisRepository)
     {
-      $this->ressourceRepository = $ressourceRepository;
-      $this->lienRessourceRepository = $lienRessourceRepository;
+        $this->ressourceRepository = $ressourceRepository;
+        $this->lienRessourceRepository = $lienRessourceRepository;
+        $this->favorisRepository = $favorisRepository;
     }
-  
-    
+
+
+
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $lesRessources = $this->ressourceRepository->all();
+
+        $validated = $request->validate([
+            'type' => 'sometimes|int',
+            'categorie' => 'sometimes|int',
+        ]);
+
+        $lesRessources = $this->ressourceRepository->allPublic();
+
+
+        $idType = null;
+        $idCategorie = null;
+
+        try {
+            $idType = $validated['type'];
+        } catch (Exception $e) {
+        }
+        try {
+            $idCategorie = $validated['categorie'];
+        } catch (Exception $e) {
+        }
+
+
+        if ($idType != null) {
+            $lesRessources = $this->ressourceRepository->AllPublicByType($idType);
+        }
+        if ($idCategorie != null) {
+            if ($idType != null) {
+                $lesRessources = $this->ressourceRepository->AllPublicByTypeAndCateg($idType, $idCategorie);
+            } else {
+                $lesRessources = $this->ressourceRepository->AllPublicByCategorie($idCategorie);
+            }
+        }
+
+
+
         return response()->json($lesRessources);
     }
+
+
+
+    public function indexUtilisateur(Request $request)
+    {
+        $authID = auth()->user()->id;
+        $result = $this->ressourceRepository->allPrivees($authID);
+        if ($request->favoris) {
+            $result = $result->merge($this->favorisRepository->whereFavoris($authID));
+        }
+        if ($request->partagees) {
+            $result = $result->merge($this->lienRessourceRepository->findRessourceByIdUser($authID));
+        }
+        if ($request->mise_de_cote) {
+            $result = $result->merge($this->favorisRepository->whereMiseDeCote($authID));
+        }
+
+        return response()->json($result);
+    }
+
 
     /**
      * Store a newly created resource in storage.
@@ -46,8 +105,9 @@ class RessourceController extends Controller
     public function store(StoreUpdateRessourceRequest $request)
     {
         $validated = $request->validated();
+        $result = array_merge($validated, ['IdUtilisateur_createur' => auth()->user()->id]);
 
-        Ressources::create($validated);
+        Ressources::create($result);
     }
 
     /**
@@ -56,13 +116,39 @@ class RessourceController extends Controller
      * @param  \App\Models\Ressources  $ressource
      * @return \Illuminate\Http\Response
      */
-    public function show(Ressources $ressource)
+    public function show(Ressources $ressource, Request $request)
     {
-        $laRessource =[
-            'ressource' => $this->ressourceRepository->one($ressource)
-          ];
-          return response()->json($laRessource);
+
+        $autorisation = false;
+        $result = ["message" => "Vous n'avez pas la permission"];
+        if ($ressource->IdEtat == 4) {
+            $autorisation = true;
+        } elseif (auth('api')->user() != null) {
+            $authID = auth('api')->user()->id;
+            $createurRessource = $ressource->Utilisateur()->get();
+            if ($createurRessource[0]['id'] == $authID) {
+                $autorisation = true;
+            } else {
+                $lesRessourcesPartages = $this->lienRessourceRepository->findRessourceByIdUser($authID);
+                foreach ($lesRessourcesPartages as $uneRessource) {
+                    if ($uneRessource['id'] == $ressource->id) {
+                        $autorisation = true;
+                    }
+                }
+            }
+        }
+
+        if ($autorisation) {
+            $result = $this->ressourceRepository->one($ressource);
+        }
+
+        // return response()->json($result);
+        return response()->json($result);
+            // return auth('api')->user();
+        ;
     }
+
+
 
     /**
      * Update the specified resource in storage.
@@ -73,10 +159,35 @@ class RessourceController extends Controller
      */
     public function update(StoreUpdateRessourceRequest $request, $id)
     {
-        $ressource = Ressources::find($id);
+        $result = ["message" => "Vous n'avez pas les droits"];
+
+        $autorisation = false;
+        $ressourceAModif = Ressources::find($id);
+        $ressource = $this->ressourceRepository->findById($id);
         $validated = $request->validated();
-        
-        $ressource->update($validated);
+        $idAuth = auth()->user()->id;
+
+        if ($this->ressourceRepository->findCreateur($ressource['id']) == $idAuth) {
+            //Si elle est privé, il faut supprimer les liens avec
+            $autorisation = true;
+        } elseif ($this->lienRessourceRepository->findCorrespRessourcesUtilisateurs($idAuth, $id)) {
+            $autorisation = true;
+        }
+
+        if ($autorisation) {
+            $ressourceAModif->update($validated);
+            try {
+                if ($validated['IdEtat'] == 1) {
+                    $lesLiensRessource = $this->lienRessourceRepository->findByIdRessource($id);
+                    foreach ($lesLiensRessource as $unLienRessource) {
+                        Jointure_ress_utilisateur::destroy($unLienRessource['id']);
+                    }
+                }
+            } catch (Exception $e) {
+            }
+            $result = ["message" => "Mise à jour effectuée"];
+        }
+        return response()->json($result);
     }
 
     /**
@@ -87,15 +198,21 @@ class RessourceController extends Controller
      */
     public function destroy($id)
     {
+        $result = ["message" => "Vous n'avez pas les droits"];
         $ressource = $this->ressourceRepository->findById($id);
-        $lesLiensRessource = $this->lienRessourceRepository->findByIdRessource($id);
-        
-        foreach($lesLiensRessource as $unLienRessource){
-            Jointure_ress_utilisateur::destroy($unLienRessource['id']);
+        if ($this->ressourceRepository->findCreateur($ressource['id']) == auth()->user()->id) {
+            $lesLiensRessource = $this->lienRessourceRepository->findByIdRessource($id);
+
+            foreach ($lesLiensRessource as $unLienRessource) {
+                Jointure_ress_utilisateur::destroy($unLienRessource['id']);
+            }
+            foreach ($ressource['commentaires'] as $unCommentaire) {
+                Commentaire::destroy($unCommentaire['id']);
+            }
+            Ressources::destroy($id);
+            $result = ["message" => "Suppression effectuée"];
+        } else {
         }
-         foreach($ressource['commentaires'] as $unCommentaire){
-            Commentaire::destroy($unCommentaire['id']);
-        }
-        Ressources::destroy($id);
+        return response()->json($result);
     }
 }
